+++ lib/apartment_acme_client/acme_client/real_client.rb
@@ -1,21 +1,23 @@
+# frozen_string_literal: true
+
 require 'acme-client'
 
 module ApartmentAcmeClient
   module AcmeClient
     class RealClient
-      def initialize(private_key:)
+      attr_reader :csr_private_key
+      def initialize(acme_client_private_key:, csr_private_key:)
         @client = Acme::Client.new(
-          private_key: private_key,
-          endpoint: server_endpoint,
+          private_key: acme_client_private_key,
+          directory: server_directory
         )
+        @csr_private_key = csr_private_key
       end
 
       def register(email)
         # If the private key is not known to the server, we need to register it for the first time.
-        registration = @client.register(contact: "mailto:#{email}")
-
-        # You may need to agree to the terms of service (that's up the to the server to require it or not but boulder does by default)
-        registration.agree_terms
+        account = @client.new_account(contact: "mailto:#{email}", terms_of_service_agreed: true)
+        Rollbar.info("New Let's Encrypt Account created with KID: #{account.kid}")
 
         true
       end
@@ -24,31 +26,35 @@ module ApartmentAcmeClient
         @client.authorize(domain: domain)
       end
 
+      def new_order(identifiers:)
+        @client.new_order(identifiers: identifiers)
+      end
+
       # Create a Certificate for our new set of domain names
       # returns that certificate
-      def request_certificate(common_name:, domains:)
+      def request_certificate(common_name:, names:, order:)
         # We're going to need a certificate signing request. If not explicitly
         # specified, the first name listed becomes the common name.
-        csr = Acme::Client::CertificateRequest.new(common_name: common_name, names: domains)
+        csr = Acme::Client::CertificateRequest.new(private_key: csr_private_key, common_name: common_name, names: names)
+        order.finalize(csr: csr)
+        while order.status == 'processing'
+          sleep(1)
+          order.reload
+        end
 
-        # We can now request a certificate. You can pass anything that returns
-        # a valid DER encoded CSR when calling to_der on it. For example an
-        # OpenSSL::X509::Request should work too.
-        certificate = @client.new_certificate(csr) # => #<Acme::Client::Certificate ....>
-
-        certificate
+        order.certificate
       end
 
       private
 
-      def server_endpoint
+      def server_directory
         # We need an ACME server to talk to, see github.com/letsencrypt/boulder
         # WARNING: This endpoint is the production endpoint, which is rate limited and will produce valid certificates.
         # You should probably use the staging endpoint for all your experimentation:
         if ApartmentAcmeClient.lets_encrypt_test_server_enabled
-          'https://acme-staging.api.letsencrypt.org/'
+          'https://acme-staging-v02.api.letsencrypt.org/directory'
         else
-          'https://acme-v01.api.letsencrypt.org/'
+          'https://acme-v02.api.letsencrypt.org/directory'
         end
       end
     end