+++ lib/apartment_acme_client/encryption.rb
@@ -1,11 +1,11 @@
+# frozen_string_literal: true
+
 require 'openssl'
 
 # Initially, the system is only accessible via subdomain.example.com
 # But, as we add more Conventions, we want to be able to access those also,
 # thus we will need:
-# - a.subdomain.example.com
-# - b.subdomain.example.com
-# - c.subdomain.example.com
+# - *.subdomain.example.com
 #
 # Also, each convention may add an "alias" for their convention, like:
 # - www.naucc.com
@@ -21,20 +21,21 @@ require 'openssl'
 #
 # Manage the encryption of the website (https).
 module ApartmentAcmeClient
-  class Encryption
+  class Encryption # rubocop:disable Metrics/ClassLength
     def initialize
       @certificate_storage = ApartmentAcmeClient::CertificateStorage::Proxy.singleton
     end
 
     # Largely based on https://github.com/unixcharles/acme-client documentation
     def register_new(email)
-      raise StandardError.new("Private key already exists") unless @certificate_storage.private_key.nil?
+      raise StandardError.new('Private key already exists') unless @certificate_storage.private_key.nil?
 
       private_key = create_private_key
 
       # Initialize the client
       new_client = ApartmentAcmeClient::AcmeClient::Proxy.singleton(
-        private_key: private_key,
+        acme_client_private_key: private_key,
+        csr_private_key: nil, # not needed for 'register' call
       )
 
       new_client.register(email)
@@ -42,22 +43,75 @@ module ApartmentAcmeClient
       @certificate_storage.save_private_key(private_key)
     end
 
-    def authorize_domains(domains)
-      successful_domains = domains.select { |domain| authorize_domain(domain) }
-      successful_domains
+    # Authorize a wildcard cert domain.
+    # to do this, we have to write to the Amazon Route53 DNS entry
+    # params:
+    #  - authorizations - a list of authorizations, which may be http or dns based (ignore the non-wildcard ones)
+    #  - wildcard_domain - the url of the wildcard's base domain (e.g. "site.example.com")
+    def authorize_domains_with_dns(authorizations, wildcard_domain:) # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
+      label = nil
+      record_type = nil
+      values = []
+
+      dns_authorizations = []
+      authorizations.each do |domain_authorization|
+        next unless domain_authorization.wildcard || domain_authorization.http.nil?
+
+        dns_authorizations << domain_authorization.dns
+      end
+
+      dns_authorizations.each do |authorization|
+        label         = "#{authorization.record_name}.#{wildcard_domain}"
+        record_type   = authorization.record_type
+        value         = authorization.record_content
+        values << value
+      end
+
+      return unless values.any?
+
+      route53 = ApartmentAcmeClient::DnsApi::Route53.new(
+        requested_domain: wildcard_domain,
+        dns_record_label: label,
+        record_type: record_type,
+        values: values
+      )
+
+      route53.write_record
+
+      check_dns = ApartmentAcmeClient::DnsApi::CheckDns.new(wildcard_domain, label)
+
+      check_dns.wait_for_present(values.first)
+
+      if check_dns.check_dns(values.first)
+        # DNS is updated, proceed with cert request
+        dns_authorizations.each do |domain_authorization|
+          domain_authorization.request_validation
+
+          10.times do
+            # may be 'pending' initially
+            break if domain_authorization.status == 'valid'
+
+            puts "Waiting for LetsEncrypt to authorize the domain"
+
+            # Wait a bit for the server to make the request, or just blink. It should be fast.
+            sleep(2)
+            domain_authorization.reload
+          end
+        end
+      else
+        # ERROR, DNS not updated in time
+        Rollbar.error("DNS Entry not found in timeout")
+      end
     end
 
-    # authorizes a domain with letsencrypt server
+    # authorizes a single domain with letsencrypt server
     # returns true on success, false otherwise.
     #
     # from https://github.com/unixcharles/acme-client/tree/master#authorize-for-domain
-    def authorize_domain(domain)
-      authorization = client.authorize(domain: domain)
+    def authorize_domain_with_http(domain_authorization)
+      challenge = domain_authorization.http
 
-      # This example is using the http-01 challenge type. Other challenges are dns-01 or tls-sni-01.
-      challenge = authorization.http01
-
-      # The http-01 method will require you to respond to a HTTP request.
+      # The http method will require you to respond to a HTTP request.
 
       # You can retrieve the challenge token
       challenge.token # => "some_token"
@@ -90,39 +144,75 @@ module ApartmentAcmeClient
       #  challenge = client.challenge_from_hash(JSON.parse(File.read('challenge')))
 
       # Once you are ready to serve the confirmation request you can proceed.
-      challenge.request_verification # => true
+      challenge.request_validation # => true
 
-      success = false
       10.times do
-        if challenge.verify_status == 'valid' # may be 'pending' initially
-          success = true
-          break
-        end
+        # may be 'pending' initially
+        break if challenge.status == 'valid'
+
+        puts "Waiting for letsencrypt to authorize the single domain"
 
         # Wait a bit for the server to make the request, or just blink. It should be fast.
-        sleep(1)
+        sleep(2)
+        challenge.reload
       end
       File.delete(full_challenge_filename)
 
-      success
+      challenge.status == 'valid'
     end
 
-    def request_certificate(common_name:, domains:)
-      client.request_certificate(common_name: common_name, domains: domains)
+    # Create an order, perform authorization for each domain, and then
+    # request the certificate.
+    # - common name is used so that there is continuity of requests over time
+    # - domains are the list of individual http-based domains to be authorized
+    # - wildcard_domain is an optional wildcard domain to be authorized via DNS Record
+    #
+    # Returns the certificate
+    def request_certificate(common_name:, domains:, wildcard_domain: nil)
+      domain_names_requested = domains
+      domain_names_requested += [wildcard_domain, "*.#{wildcard_domain}"] if wildcard_domain.present?
+      order = client.new_order(identifiers: domain_names_requested)
+
+      # Do the HTTP authorizations
+      order.authorizations.each do |authorization|
+        encryptor.send(:authorize_domain_with_http, authorization) unless authorization.wildcard || authorization.http.nil?
+      end
+      # Do the DNS (wildcard) authorizations
+      authorize_domains_with_dns(order.authorizations, wildcard_domain: wildcard_domain)
+
+      client.request_certificate(common_name: common_name, names: domain_names_requested, order: order)
+    end
+
+    # for use in order to store this on the machine for NGINX use
+    def csr_private_key_string
+      csr_private_key.to_s
     end
 
     private
 
     def client
       @client ||= ApartmentAcmeClient::AcmeClient::Proxy.singleton(
-        private_key: private_key,
+        acme_client_private_key: acme_client_private_key,
+        csr_private_key: csr_private_key,
       )
     end
 
     # Returns a private key
-    def private_key
+    def acme_client_private_key
       private_key = @certificate_storage.private_key
       return nil unless private_key
+
+      OpenSSL::PKey::RSA.new(private_key)
+    end
+
+    def csr_private_key
+      private_key = @certificate_storage.csr_private_key
+
+      # create a new private key if one is not found
+      if private_key.nil?
+        private_key = create_private_key
+        @certificate_storage.save_csr_private_key(private_key)
+      end
 
       OpenSSL::PKey::RSA.new(private_key)
     end