+++ spec/apartment_acme_client/stubs/fake_certificate_storage.rb
@@ -1,6 +1,9 @@
 module Stubs
   class FakeCertificateStorage
-    def store_certificate(certificate)
+    def store_certificate_string(certificate)
+    end
+
+    def store_csr_private_key_string(private_key)
     end
 
     def cert_exists?
@@ -9,7 +12,14 @@ module Stubs
     def private_key
     end
 
+    def csr_private_key
+    end
+
     def save_private_key(private_key)
+      true
+    end
+
+    def save_csr_private_key(private_key)
       true
     end
   end